backend/requirements.txt
flask
flask-cors
transformers
torch
numpy
gunicorn
python-dotenv
python-docx